Coalesce transition effect resets

+++ src/cgame/cg_local.h
@@ -988,8 +988,8 @@ typedef struct {
 
   // blend blobs
   viewDamage_t viewDamage[MAX_VIEWDAMAGE];
-  float damageTime; // last time any kind of damage was recieved
-  int damageIndex;  // slot that was filled in
+  int damageTime;  // last time any kind of damage was recieved
+  int damageIndex; // slot that was filled in
   float damageX, damageY, damageValue;
 
   int grenLastTime;
